# Optimize JsonReader and JsonParser, and update design for cleaner lifetime semantics