--- a/src/System.Text.JsonLab/System/Text/Json/CustomDb.cs
+++ b/src/System.Text.JsonLab/System/Text/Json/CustomDb.cs
@@ -83,9 +83,9 @@ namespace System.Text.JsonLab
         {
             Debug.Assert(index >= 0 && index <= Span.Length - DbRow.Size);
             Debug.Assert(numberOfRows >= 1 && numberOfRows <= 0x0FFFFFFF);
-            byte previous = (byte)(Span[index + DbRow.Size - 1] >> 4);
+            byte previous = (byte)(Span[index + 11] >> 4);
             int value = (previous << 28) | numberOfRows;
-            MemoryMarshal.Write(Span.Slice(index + DbRow.Size - 4), ref value);
+            MemoryMarshal.Write(Span.Slice(index + 8), ref value);
         }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
@@ -93,9 +93,18 @@ namespace System.Text.JsonLab
         public void SetHasChildren(int index)
         {
             Debug.Assert(index >= 0 && index <= Span.Length - DbRow.Size);
-            index += DbRow.Size - 1;
+            index += 11;
             byte previous = Span[index];
             Span[index] = (byte)(1 << 7 | previous);
+        }
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        //TODO: Resolve little endian/big endian issue
+        public void SetArrayUniformLength(int index, int uniformLength)
+        {
+            Debug.Assert(index >= 0 && index <= Span.Length - DbRow.Size);
+            Debug.Assert(uniformLength >= 1 && uniformLength <= 0x0FFFFFFF);
+            MemoryMarshal.Write(Span.Slice(index + 12), ref uniformLength);
         }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
@@ -186,11 +195,11 @@ namespace System.Text.JsonLab
         public string PrintDatabase()
         {
             StringBuilder sb = new StringBuilder();
-            sb.Append(nameof(DbRow.Location) + "\t" + nameof(DbRow.SizeOrLength) + "\t" + nameof(DbRow.JsonType) + "\t" + nameof(DbRow.HasChildren) + "\t" + nameof(DbRow.NumberOfRows) + "\r\n");
+            sb.Append(nameof(DbRow.Location) + "\t" + nameof(DbRow.SizeOrLength) + "\t" + nameof(DbRow.JsonType) + "\t" + nameof(DbRow.HasChildren) + "\t" + nameof(DbRow.NumberOfRows) + "\t" + nameof(DbRow.ArrayUniformLength) + "\r\n");
             for (int i = 0; i < Span.Length; i += DbRow.Size)
             {
                 DbRow record = MemoryMarshal.Read<DbRow>(Span.Slice(i));
-                sb.Append(record.Location + "\t" + record.SizeOrLength + "\t" + record.JsonType + "\t" + record.HasChildren + "\t" + record.NumberOfRows + "\r\n");
+                sb.Append(record.Location + "\t" + record.SizeOrLength + "\t" + record.JsonType + "\t" + record.HasChildren + "\t" + record.NumberOfRows + "\t" + record.ArrayUniformLength + "\r\n");
             }
             return sb.ToString();
         }