… src/System.Text.JsonLab/System/Text/Json/JsonObject.cs
… src/System.Text.JsonLab/System/Text/Json/JsonParser.cs
… src/System.Text.JsonLab/System/Text/Json/JsonReader.cs
… src/System.Text.JsonLab/System/Text/Json/JsonThrowHelper.cs
… src/System.Text.JsonLab/System/Text/Json/StackRow.cs
… tests/Benchmarks/System.Text.JsonLab/JsonParserChangeEntryPerf.cs
… tests/Benchmarks/System.Text.JsonLab/JsonParserEnumeratePerf.cs
… tests/Benchmarks/System.Text.JsonLab/JsonParserPerf.cs
… tests/Benchmarks/System.Text.JsonLab/JsonReaderPerf.cs
… tests/System.Text.JsonLab.Tests/JsonObjectTests.cs
… tests/System.Text.JsonLab.Tests/JsonReaderTests.cs
… tests/System.Text.JsonLab.Tests/Resources/TestJson.resx