- Add SkipPastAny overloads.
- Add SkipPastSlow which advances 1 at a time.
- Add back static JsonObject.Parse method.
- Break out the db functionality from JsonParser into a new CustomDb type.
- Add a union column in DbRow that stores more info during parsing.
- Use a static byte array for the initial stack space.
- Add JsonParserEnumeratePerf benchmark. Experiments with attempting amortized O(n) enumeration of arrays was only beneficial for large json objects. We should revisit this change.
- Add compact (no whitespace/formatting) json payloads to benchmarks.
- Add parser benchmark mode where we rebuild the json string from the parsing loop.
- Change JsonReader implementation to optimize the zero-space cases.

``` ini
BenchmarkDotNet=v0.10.14.683-nightly, OS=Windows 10.0.17713
Intel Core i7-6700 CPU 3.40GHz (Skylake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=3.0.100-alpha1-20180720-2
  [Host]     : .NET Core 2.1.2 (CoreCLR 4.6.26628.05, CoreFX 4.6.26629.01), 64bit RyuJIT
  DefaultJob : .NET Core 2.1.2 (CoreCLR 4.6.26628.05, CoreFX 4.6.26629.01), 64bit RyuJIT
```

**Before:**

|                 Method | IsDataCompact | OnlyParse |   TestCase |           Mean |         Error |        StdDev |  Gen 0 | Allocated |
|----------------------- |-------------- |---------- |----------- |---------------:|--------------:|--------------:|-------:|----------:|
| **ParseSystemTextJsonLab** |          **True** |     **False** |  **BasicJson** |     **4,349.4 ns** |     **48.266 ns** |     **42.786 ns** | **0.2594** |    **1088 B** |
| **ParseSystemTextJsonLab** |          **True** |     **False** | **HelloWorld** |       **688.9 ns** |     **12.978 ns** |     **12.746 ns** | **0.0191** |      **88 B** |
| **ParseSystemTextJsonLab** |          **True** |     **False** |   **Json400B** |     **7,522.4 ns** |    **146.124 ns** |    **129.535 ns** | **0.5341** |    **2296 B** |
| **ParseSystemTextJsonLab** |          **True** |     **False** |  **Json400KB** |             **NA** |            **NA** |            **NA** |    **N/A** |       **N/A** |
| **ParseSystemTextJsonLab** |          **True** |     **False** |   **Json40KB** |             **NA** |            **NA** |            **NA** |    **N/A** |       **N/A** |
| **ParseSystemTextJsonLab** |          **True** |     **False** |    **Json4KB** |             **NA** |            **NA** |            **NA** |    **N/A** |       **N/A** |
| **ParseSystemTextJsonLab** |          **True** |      **True** |  **BasicJson** |     **1,758.5 ns** |     **13.974 ns** |     **13.071 ns** |      **-** |       **0 B** |
| **ParseSystemTextJsonLab** |          **True** |      **True** | **HelloWorld** |       **424.7 ns** |      **2.690 ns** |      **2.516 ns** |      **-** |       **0 B** |
| **ParseSystemTextJsonLab** |          **True** |      **True** |   **Json400B** |     **2,351.3 ns** |     **32.048 ns** |     **26.761 ns** |      **-** |       **0 B** |
| **ParseSystemTextJsonLab** |          **True** |      **True** |  **Json400KB** | **1,875,418.0 ns** | **13,582.975 ns** | **12,705.522 ns** |      **-** |       **0 B** |
| **ParseSystemTextJsonLab** |          **True** |      **True** |   **Json40KB** |   **181,970.2 ns** |  **2,600.134 ns** |  **2,304.951 ns** |      **-** |       **0 B** |
| **ParseSystemTextJsonLab** |          **True** |      **True** |    **Json4KB** |    **17,087.7 ns** |    **255.761 ns** |    **239.239 ns** |      **-** |       **0 B** |

**After:**

|                 Method | IsDataCompact | OnlyParse |   TestCase |           Mean |          Error |         StdDev | Scaled |    Gen 0 |    Gen 1 |   Gen 2 | Allocated |
|----------------------- |-------------- |---------- |----------- |---------------:|---------------:|---------------:|-------:|---------:|---------:|--------:|----------:|
| **ParseSystemTextJsonLab** |          **True** |     **False** |  **BasicJson** |     **3,814.5 ns** |      **76.070 ns** |     **170.141 ns** |   **1.00** |   **0.2594** |        **-** |       **-** |    **1088 B** |
|                        |               |           |            |                |                |                |        |          |          |         |           |
| **ParseSystemTextJsonLab** |          **True** |     **False** | **HelloWorld** |       **575.0 ns** |      **11.254 ns** |      **17.186 ns** |   **1.00** |   **0.0200** |        **-** |       **-** |      **88 B** |
|                        |               |           |            |                |                |                |        |          |          |         |           |
| **ParseSystemTextJsonLab** |          **True** |     **False** |   **Json400B** |     **7,259.4 ns** |     **192.590 ns** |     **523.955 ns** |   **1.00** |   **0.5417** |        **-** |       **-** |    **2296 B** |
|                        |               |           |            |                |                |                |        |          |          |         |           |
| **ParseSystemTextJsonLab** |          **True** |     **False** |  **Json400KB** | **9,682,912.5 ns** | **252,709.150 ns** | **733,155.066 ns** |   **1.00** | **296.8750** | **203.1250** | **93.7500** | **1545428 B** |
|                        |               |           |            |                |                |                |        |          |          |         |           |
| **ParseSystemTextJsonLab** |          **True** |     **False** |   **Json40KB** |   **469,846.8 ns** |   **9,346.559 ns** |  **14,273.183 ns** |   **1.00** |  **36.6211** |        **-** |       **-** |  **156656 B** |
|                        |               |           |            |                |                |                |        |          |          |         |           |
| **ParseSystemTextJsonLab** |          **True** |     **False** |    **Json4KB** |    **45,779.1 ns** |     **910.711 ns** |     **807.322 ns** |   **1.00** |   **4.3945** |        **-** |       **-** |   **18440 B** |
|                        |               |           |            |                |                |                |        |          |          |         |           |
| **ParseSystemTextJsonLab** |          **True** |      **True** |  **BasicJson** |     **1,041.7 ns** |      **18.954 ns** |      **15.827 ns** |   **1.00** |        **-** |        **-** |       **-** |       **0 B** |
|                        |               |           |            |                |                |                |        |          |          |         |           |
| **ParseSystemTextJsonLab** |          **True** |      **True** | **HelloWorld** |       **281.0 ns** |       **5.549 ns** |       **5.937 ns** |   **1.00** |        **-** |        **-** |       **-** |       **0 B** |
|                        |               |           |            |                |                |                |        |          |          |         |           |
| **ParseSystemTextJsonLab** |          **True** |      **True** |   **Json400B** |     **1,501.3 ns** |      **41.962 ns** |      **35.040 ns** |   **1.00** |        **-** |        **-** |       **-** |       **0 B** |
|                        |               |           |            |                |                |                |        |          |          |         |           |
| **ParseSystemTextJsonLab** |          **True** |      **True** |  **Json400KB** | **1,267,946.4 ns** |  **25,046.191 ns** |  **29,815.708 ns** |   **1.00** |        **-** |        **-** |       **-** |       **0 B** |
|                        |               |           |            |                |                |                |        |          |          |         |           |
| **ParseSystemTextJsonLab** |          **True** |      **True** |   **Json40KB** |   **121,033.9 ns** |   **2,396.071 ns** |   **3,279.769 ns** |   **1.00** |        **-** |        **-** |       **-** |       **0 B** |
|                        |               |           |            |                |                |                |        |          |          |         |           |
| **ParseSystemTextJsonLab** |          **True** |      **True** |    **Json4KB** |    **10,486.9 ns** |     **200.862 ns** |     **268.145 ns** |   **1.00** |        **-** |        **-** |       **-** |       **0 B** |

**Before:**

|                                            Method | IsDataCompact |   TestCase |           Mean |         Error |        StdDev |         Median |  Gen 0 | Allocated |
|-------------------------------------------------- |-------------- |----------- |---------------:|--------------:|--------------:|---------------:|-------:|----------:|
|              **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |  **BasicJson** |     **3,544.9 ns** |    **425.366 ns** |  **1,240.811 ns** |     **3,200.0 ns** |      **-** |       **0 B** |
| ReaderSystemTextJsonLabMultiSpanSequenceEmptyLoop |          True |  BasicJson |     1,434.9 ns |     10.236 ns |      9.575 ns |     1,436.2 ns | 0.0076 |      40 B |
|              **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** | **HelloWorld** |       **125.8 ns** |      **4.125 ns** |      **3.657 ns** |       **124.8 ns** |      **-** |       **0 B** |
| ReaderSystemTextJsonLabMultiSpanSequenceEmptyLoop |          True | HelloWorld |       498.2 ns |      4.788 ns |      4.479 ns |       499.0 ns | 0.0086 |      40 B |
|              **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |   **Json400B** |     **1,226.3 ns** |     **11.142 ns** |     **10.422 ns** |     **1,226.0 ns** |      **-** |       **0 B** |
| ReaderSystemTextJsonLabMultiSpanSequenceEmptyLoop |          True |   Json400B |     1,715.0 ns |     25.443 ns |     23.800 ns |     1,712.6 ns |      - |       0 B |
|              **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |  **Json400KB** | **1,127,673.7 ns** |  **8,321.665 ns** |  **7,376.938 ns** | **1,127,924.8 ns** |      **-** |       **0 B** |
| ReaderSystemTextJsonLabMultiSpanSequenceEmptyLoop |          True |  Json400KB | 1,410,088.7 ns | 14,474.088 ns | 11,300.415 ns | 1,411,198.3 ns |      - |     440 B |
|              **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |   **Json40KB** |   **111,835.2 ns** |  **1,053.686 ns** |    **985.618 ns** |   **111,523.0 ns** |      **-** |       **0 B** |
| ReaderSystemTextJsonLabMultiSpanSequenceEmptyLoop |          True |   Json40KB |   140,418.5 ns |  2,771.333 ns |  4,062.184 ns |   138,781.6 ns |      - |      40 B |
|              **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |    **Json4KB** |     **9,467.5 ns** |     **94.473 ns** |     **88.370 ns** |     **9,482.8 ns** |      **-** |       **0 B** |
| ReaderSystemTextJsonLabMultiSpanSequenceEmptyLoop |          True |    Json4KB |    12,664.7 ns |    124.454 ns |    116.414 ns |    12,641.6 ns | 0.0916 |     416 B |

**After:**

|                                            Method | IsDataCompact |   TestCase |            Mean |           Error |          StdDev |  Gen 0 | Allocated |
|-------------------------------------------------- |-------------- |----------- |----------------:|----------------:|----------------:|-------:|----------:|
|              **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |  **BasicJson** |       **649.97 ns** |       **4.1203 ns** |       **3.6525 ns** |      **-** |       **0 B** |
| ReaderSystemTextJsonLabMultiSpanSequenceEmptyLoop |          True |  BasicJson |     1,390.76 ns |      14.8609 ns |      13.1738 ns | 0.0076 |      40 B |
|              **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** | **HelloWorld** |        **94.59 ns** |       **0.8219 ns** |       **0.7286 ns** |      **-** |       **0 B** |
| ReaderSystemTextJsonLabMultiSpanSequenceEmptyLoop |          True | HelloWorld |       496.26 ns |       5.3908 ns |       4.5016 ns | 0.0086 |      40 B |
|              **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |   **Json400B** |       **986.31 ns** |      **12.7486 ns** |      **11.9251 ns** |      **-** |       **0 B** |
| ReaderSystemTextJsonLabMultiSpanSequenceEmptyLoop |          True |   Json400B |     1,699.47 ns |      19.0710 ns |      16.9059 ns |      - |       0 B |
|              **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |  **Json400KB** |   **953,232.28 ns** |  **18,279.9188 ns** |  **21,051.1798 ns** |      **-** |       **0 B** |
| ReaderSystemTextJsonLabMultiSpanSequenceEmptyLoop |          True |  Json400KB | 1,477,274.90 ns | 201,523.3781 ns | 188,505.0746 ns |      - |     440 B |
|              **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |   **Json40KB** |    **88,480.21 ns** |   **1,254.3162 ns** |   **1,111.9185 ns** |      **-** |       **0 B** |
| ReaderSystemTextJsonLabMultiSpanSequenceEmptyLoop |          True |   Json40KB |   134,772.32 ns |   1,230.6454 ns |   1,151.1464 ns |      - |      40 B |
|              **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |    **Json4KB** |     **7,493.69 ns** |     **147.1333 ns** |     **279.9363 ns** |      **-** |       **0 B** |
| ReaderSystemTextJsonLabMultiSpanSequenceEmptyLoop |          True |    Json4KB |    12,349.82 ns |     211.9233 ns |     176.9656 ns | 0.0916 |     416 B |